Make dxAccordion SSR-compatible
